backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
fonttools==4.59.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.3.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
threadpoolctl==3.6.0
tweepy==4.15.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
wordcloud==1.9.4
